asignaturas endpoint, actualización mapeos

+++ src/main/java/com/niglesiasm/eduapp/repository/curso/CursoDao.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CursoDao extends JpaRepository<Curso, Long> {
+public interface CursoDao extends JpaRepository<Curso, Integer> {
 
 }
